requirements.txt
tensorflow==2.15.0
fastapi==0.51.0
uvicorn
scikit-learn
pandas
Pillow
python-multipart
PyJWT
python-decouple
gcsfs
astroid==2.3.3
Click==7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pydantic
pylint==2.4.4
six==1.14.0
starlette==0.13.2
wrapt==1.11.2
google-cloud-firestore
gunicorn
python-jose